feat: improve way floting components position is updated
